로또 넘버 객체
