--- src/main/java/lotto/domain/LottoCreatingStrategy.java
@@ -1,5 +1,0 @@
-package lotto.domain;
-
-public interface LottoCreatingStrategy {
-    Lotto create();
-}
